requirements.txt
pandas
openpyxl
plotly
requests
pyyaml